Admin completo

+++ src/main/java/mx/com/web/controller/RegistroController.java
@@ -49,16 +49,38 @@ public class RegistroController {
 	private GrupoService grupoService;
 	
 	@RequestMapping(value="/registro", method = RequestMethod.GET)
-	public ModelAndView regitro(){
+	public ModelAndView regitro(@RequestParam(value="aux", required = false) String aux){
 		ModelAndView model = new ModelAndView();
 		model.setViewName("registro");
 		model.addObject("grupos", grupoService.getAllGrupos());
+		ObjectMapper mapper = new ObjectMapper();
+		if(aux.equals("0"))
+			model.addObject("alumno", "{}");
+		else{
+			try {
+				model.addObject("alumno", mapper.writeValueAsString(personaService.getPersonasById(Integer.parseInt(aux))));
+			} catch (NumberFormatException | JsonProcessingException e) {
+				log.error(e.toString());
+			}
+		}
 		return model;
 	}
 		
 	@RequestMapping(value="/admin/registro", method = RequestMethod.GET)
-	public String registroprofesor(){
-		return "registroprofesor";
+	public ModelAndView registroprofesor(@RequestParam(value="aux", required = false) String aux){
+		ModelAndView model = new ModelAndView();
+		model.setViewName("registroprofesor");
+		ObjectMapper mapper = new ObjectMapper();
+		if(aux.equals("0"))
+			model.addObject("profesor", "{}");
+		else{
+			try {
+				model.addObject("profesor", mapper.writeValueAsString(personaService.getPersonasById(Integer.parseInt(aux))));
+			} catch (NumberFormatException | JsonProcessingException e) {
+				log.error(e.toString());
+			}
+		}
+		return model;
 	}
 		
 	@RequestMapping(value="/registrar", method = RequestMethod.POST)
@@ -68,6 +90,7 @@ public class RegistroController {
 		String estatus = "OK";
 		try {
 			alu = mapper.readValue(registro,Persona.class);
+			if(alu.getId() == 0){
 			Users users = new Users();
 			users.setUsername(alu.getUsername());
 			users.setPassword(pass.Generate("123456"));
@@ -81,7 +104,10 @@ public class RegistroController {
 					estatus = "NOOK";
 				}
 			}
-			
+			}else{
+				if(!personaService.UpdatePersona(alu))
+					estatus="NOOK";
+			}
 		} catch (IOException e) {
 			log.error(e.toString());
 		}
@@ -98,21 +124,25 @@ public class RegistroController {
 		String estatus = "OK";
 		try {
 			pro = mapper.readValue(registro,Persona.class);
-			pro.setGrupo(1);
-			Users users = new Users();
-			users.setUsername(pro.getUsername());
-			users.setPassword(pass.Generate(pro.getContrasena()));
-			users.setEnabled(1);
-			UserRoles userRole = new UserRoles();
-			userRole.setRole(ROLE_PROFE);
-			userRole.setUsername(pro.getUsername());
-			if(adminService.InsertUser(users, userRole)){
-				if(!personaService.InsertPersona(pro)){
-					adminService.DeleteUser(users, userRole);
-					estatus = "NOOK";
+			if(pro.getId()==0){
+				pro.setGrupo(1);
+				Users users = new Users();
+				users.setUsername(pro.getUsername());
+				users.setPassword(pass.Generate(pro.getContrasena()));
+				users.setEnabled(1);
+				UserRoles userRole = new UserRoles();
+				userRole.setRole(ROLE_PROFE);
+				userRole.setUsername(pro.getUsername());
+				if(adminService.InsertUser(users, userRole)){
+					if(!personaService.InsertPersona(pro)){
+						adminService.DeleteUser(users, userRole);
+						estatus = "NOOK";
+					}
 				}
+			}else{
+				if(!personaService.UpdatePersona(pro))
+					estatus="NOOK";
 			}
-			
 		} catch (IOException e) {
 			log.error(e.toString());
 		}
@@ -143,6 +173,32 @@ public class RegistroController {
 		}else{
 			model.addObject("respuesta", "[]");
 		}
+		return model;
+	}
+	
+	@RequestMapping(value="/admin/profesores", method = RequestMethod.GET)
+	public ModelAndView getProfesoresLista(){
+		ModelAndView model = new ModelAndView();
+		ObjectMapper mapper = new ObjectMapper();
+		model.setViewName("profesores");
+		try {
+			model.addObject("profesores", mapper.writeValueAsString(personaService.getPersonas(ROLE_PROFE)));
+		} catch (JsonProcessingException e) {
+			log.error(e.toString());
+		}		
+		return model;
+	}
+	
+	@RequestMapping(value="/admin/alumnos", method = RequestMethod.GET)
+	public ModelAndView getAlumnoLista(){
+		ModelAndView model = new ModelAndView();
+		ObjectMapper mapper = new ObjectMapper();
+		model.setViewName("alumnos");
+		try {
+			model.addObject("alumnos", mapper.writeValueAsString(personaService.getPersonas(ROLE_ALUMNO)));
+		} catch (JsonProcessingException e) {
+			log.error(e.toString());
+		}		
 		return model;
 	}
 }
